pyproject.toml
[project]
name = "tem-seg"
version = "0.3.1.dev0"
description = "Codebase for the analysis of mitochondria morphology using semantic segmentation on TEM images."
readme = "README.md"
authors = [
    { name = "Argenis Arriojas", email = "arriojasmaldonado001@umb.edu" },
]
requires-python = ">=3.12,<3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.1",
    "numpy<2.2",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pydantic-settings>=2.10.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "seaborn>=0.13.2",
    "tensorflow>=2.16.0",
    "toml>=0.10.2",
    "typer>=0.16.0",
]

[project.optional-dependencies]
cuda = ["tensorflow[and-cuda]"]
gpu-mac = ["tensorflow-metal"]

[project.scripts]
tem-seg = "tem_seg.cmd:app"

[build-system]
requires = ["flit_core>=3.2,<4"]
build-backend = "flit_core.buildapi"

[dependency-groups]
dev = ["bump-my-version>=1.1.4", "pytest>=8.4.1", "ruff>=0.12.4"]
